requirements.txt
Django==5.2.4
django-allauth==65.10.0
requests==2.32.4
PyJWT==2.10.1
cryptography==45.0.5
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0